refactor: extract useTodayActions hook from TodayActionList
Move mutation state, action handlers, and keyboard navigation (j/k/arrows)
into useTodayActions(flatItems, onTickerSelect). Component retains only
data loading, derived maps, flatItems, and rendering.
