+++ web-ui/src/components/domain/today/TodayActionList.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { RefreshCw } from 'lucide-react';
 import { cn } from '@/utils/cn';
 import { t } from '@/i18n/t';
@@ -9,12 +9,9 @@ import { useDailyReview } from '@/features/dailyReview/api';
 import { parseUniverseFromStorage, SCREENER_UNIVERSE_STORAGE_KEY } from '@/features/screener/universeStorage';
 import {
   usePositions,
-  useUpdateStopMutation,
-  useClosePositionMutation,
   useOpenPositionsIntelligence,
-  usePartialClosePositionMutation,
 } from '@/features/portfolio/hooks';
-import type { ClosePositionRequest, PartialCloseRequest, Position, UpdateStopRequest } from '@/features/portfolio/types';
+import { useTodayActions } from './useTodayActions';
 import {
   CloseItem,
   UpdateStopItem,
@@ -69,68 +66,6 @@ export default function TodayActionList({ onTickerSelect }: TodayActionListProps
     [openPositionsQuery.data],
   );
 
-  const [doneIds, setDoneIds] = useState<Set<string>>(() => new Set());
-
-  const acceptStopMutation = useUpdateStopMutation();
-  const [acceptedStops, setAcceptedStops] = useState<Set<string>>(new Set());
-
-  const handleAcceptStop = useCallback(
-    (positionId: string, stopSuggested: number, reason: string) => {
-      acceptStopMutation.mutate(
-        { positionId, request: { newStop: stopSuggested, reason } },
-        { onSuccess: () => setAcceptedStops((prev) => new Set([...prev, positionId])) },
-      );
-    },
-    [acceptStopMutation],
-  );
-
-  const [updateStopTarget, setUpdateStopTarget] = useState<Position | null>(null);
-  const [closeTarget, setCloseTarget] = useState<Position | null>(null);
-  const [trimTarget, setTrimTarget] = useState<Position | null>(null);
-
-  const updateStopMutation = useUpdateStopMutation();
-  const closePositionMutation = useClosePositionMutation();
-  const partialCloseMutation = usePartialClosePositionMutation();
-
-  const handleUpdateStop = useCallback((position: Position, req: UpdateStopRequest) => {
-    updateStopMutation.mutate(
-      { positionId: position.positionId!, request: req },
-      {
-        onSuccess: () => {
-          setUpdateStopTarget(null);
-          setDoneIds((prev) => new Set([...prev, position.positionId!]));
-        },
-      },
-    );
-  }, [updateStopMutation]);
-
-  const handleClosePosition = useCallback((position: Position, req: ClosePositionRequest) => {
-    closePositionMutation.mutate(
-      { positionId: position.positionId!, request: req },
-      {
-        onSuccess: () => {
-          setCloseTarget(null);
-          setDoneIds((prev) => new Set([...prev, position.positionId!]));
-        },
-      },
-    );
-  }, [closePositionMutation]);
-
-  const handlePartialClose = useCallback((position: Position, req: PartialCloseRequest) => {
-    partialCloseMutation.mutate(
-      { positionId: position.positionId!, request: req },
-      {
-        onSuccess: () => {
-          setTrimTarget(null);
-        },
-      },
-    );
-  }, [partialCloseMutation]);
-
-  const [holdExpanded, setHoldExpanded] = useState(false);
-
-  const [focusedIndex, setFocusedIndex] = useState(-1);
-
   const flatItems = useMemo(
     () => [
       ...(review?.watchlistNearTrigger.map((i) => ({ ticker: i.ticker, id: `watch-${i.ticker}` })) ?? []),
@@ -145,36 +80,28 @@ export default function TodayActionList({ onTickerSelect }: TodayActionListProps
     [review],
   );
 
-  const handleItemClick = useCallback((ticker: string) => {
-    setFocusedIndex((prev) => {
-      const idx = flatItems.findIndex((fi) => fi.ticker === ticker);
-      return idx !== -1 ? idx : prev;
-    });
-    onTickerSelect(ticker);
-  }, [flatItems, onTickerSelect]);
+  const {
+    doneIds,
+    acceptedStops,
+    acceptStopMutation,
+    updateStopMutation,
+    closePositionMutation,
+    partialCloseMutation,
+    updateStopTarget,
+    setUpdateStopTarget,
+    closeTarget,
+    setCloseTarget,
+    trimTarget,
+    setTrimTarget,
+    focusedIndex,
+    handleAcceptStop,
+    handleUpdateStop,
+    handleClosePosition,
+    handlePartialClose,
+    handleItemClick,
+  } = useTodayActions(flatItems, onTickerSelect);
 
-  useEffect(() => {
-    const handler = (e: KeyboardEvent) => {
-      if (['INPUT', 'TEXTAREA', 'SELECT'].includes((e.target as HTMLElement)?.tagName)) return;
-      if (e.key === 'j' || e.key === 'ArrowDown') {
-        e.preventDefault();
-        setFocusedIndex((i) => {
-          const next = Math.min(i + 1, flatItems.length - 1);
-          if (flatItems[next]) onTickerSelect(flatItems[next].ticker);
-          return next;
-        });
-      } else if (e.key === 'k' || e.key === 'ArrowUp') {
-        e.preventDefault();
-        setFocusedIndex((i) => {
-          const prev = Math.max(i - 1, 0);
-          if (flatItems[prev]) onTickerSelect(flatItems[prev].ticker);
-          return prev;
-        });
-      }
-    };
-    window.addEventListener('keydown', handler);
-    return () => window.removeEventListener('keydown', handler);
-  }, [flatItems, onTickerSelect]);
+  const [holdExpanded, setHoldExpanded] = useState(false);
 
   const requiresActionCount =
     (review?.positionsClose.length ?? 0) + (review?.positionsUpdateStop.length ?? 0);
